requirements.txt
numpy
scipy
noise
Pillow
streamlit
scikit-image